# lib/api-zod/src/generated/types/partyRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PartyRole = (typeof PartyRole)[keyof typeof PartyRole];

export const PartyRole = {
  ADMIN: "ADMIN",
  UNDERWRITER: "UNDERWRITER",
  CSA: "CSA",
  AGENT: "AGENT",
  EMPLOYER: "EMPLOYER",
  CARRIER: "CARRIER",
  PEO: "PEO",
  VENDOR: "VENDOR",
} as const;