# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartyRole } from "./partyRole";

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  role: PartyRole;
  orgId?: string;
}